Added disqus, ads, categories and removed test posts.
Playing around with the CSS, have a lot of work to do
regarding the SCSS, mainly learning it. Testing out Google
AdSense and Disqus commenting, will probably be removed in
the future for something open source. Added a category plug-in
to help organize posts in the future. Added new favicon.